refactor: add module structure
