+++ index.js
@@ -1,12 +1,12 @@
 import { argv } from 'process';
 import os from 'os';
-import handleOs from './src/os.js';
-import calculateHash from './src/hash.js';
-import compressFile from './src/compress.js';
-import decompressFile from './src/decompress.js';
-import goUp from './src/goUp.js';
-import changeDir from './src/cd.js';
-import list from './src/list.js';
+import handleOs from './src/os/os.js';
+import calculateHash from './src/hash/hash.js';
+import compressFile from './src/archiving/compress.js';
+import decompressFile from './src/archiving/decompress.js';
+import goUp from './src/navigation/goUp.js';
+import changeDir from './src/navigation/cd.js';
+import list from './src/navigation/list.js';
 import readFile from './src/fs/readFile.js';
 import createFile from './src/fs/createFile.js';
 import renameFile from './src/fs/renameFile.js';
