+++ src/archiving/compress.js
