+++ src/archiving/decompress.js
