+++ src/hash/hash.js
